+++ app/features/search/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from "./use-search-movie";
